Throw a GradleException if .git directory is missing
`GitPropertiesPluginTests#testGenerateWithMissingGitRepoShouldFail`
incorrectly asserted that every exception would be a
`RepositoryNotFoundException` while in reality the exception was about
`Assert.fail()` not being imported and no such method could be found.

While fixing the test, I learned that JGit wouldn't throw a
`RepositoryNotFoundException` in every case but happily walk up the
directory tree until it found a `.git` directory with
`BaseRepositoryBuilder#findGitDir()` (see [1], [2]).

Instead of relying on this behavior,
`GenerateGitPropertiesTask#getGeneratedProperties()` will now throw a
`GradleException` if `dotGitDirectory` is `null` to get the expected
behavior.

[1]: https://archive.eclipse.org/jgit/site/3.7.0.201502260915-r/org.eclipse.jgit/apidocs/org/eclipse/jgit/lib/BaseRepositoryBuilder.html#findGitDir()
[2]: https://archive.eclipse.org/jgit/site/3.7.0.201502260915-r/org.eclipse.jgit/apidocs/org/eclipse/jgit/lib/BaseRepositoryBuilder.html#findGitDir(java.io.File)